add diff mode via 'Diff' command
This just highlights git diffs if you execute Diff as a command. This
feature may be removed if it's too annoying (i.e., if it starts to become
remniscent of syntax highlighting).

The implementation of this should be more relaxed to allow nested
highlighters; that way a user can do whatever they want with the
colors without inflicting trauma on the rigid color scheme.