a: v0.8.0: make window creation lazy, sync window deletions
Add an atomic delete counter to keep track of window deletions. The
writer in oscmd.go checks the counter per-write. If out of phase, we
regrab or recreate the window associated with the running command,
preventing a potential process crash.

The Del function returns the io.Closer associated with the buffer and
frame, by calling Close after incrementing the delete counter, we prevent
data races. (A historical first in this program, but more improvements will
follow suit)

fixes as/a#167 and fixes as/a#168.